Move aggregate functions to separate files
