+++ src/function/aggregate_function/count.cpp
@@ -1,0 +1,21 @@
+#include "function/aggregate_function/count.hpp"
+#include "common/exception.hpp"
+#include "common/types/null_value.hpp"
+#include "common/vector_operations/vector_operations.hpp"
+
+using namespace std;
+
+namespace duckdb {
+
+void count_update(Vector inputs[], index_t input_count, Vector &result ) {
+	assert(input_count == 1 );
+	VectorOperations::Scatter::AddOne(inputs[0], result);
+}
+
+void count_simple_update(Vector inputs[], index_t input_count, Value& result) {
+	assert(input_count == 1 );
+	Value count = VectorOperations::Count(inputs[0]);
+	result = result + count;
+}
+
+} // namespace duckdb
