+++ src/include/function/aggregate_function/avg.hpp
@@ -1,7 +1,7 @@
 //===----------------------------------------------------------------------===//
 //                         DuckDB
 //
-// function/aggregate_function/algebraic.hpp
+// function/aggregate_function/avg.hpp
 //
 //
 //===----------------------------------------------------------------------===//
