+++ src/include/function/aggregate_function/list.hpp
@@ -1,2 +1,8 @@
-#include "function/aggregate_function/algebraic.hpp"
-#include "function/aggregate_function/distributive.hpp"
+#include "function/aggregate_function/avg.hpp"
+#include "function/aggregate_function/count_star.hpp"
+#include "function/aggregate_function/count.hpp"
+#include "function/aggregate_function/first.hpp"
+#include "function/aggregate_function/max.hpp"
+#include "function/aggregate_function/min.hpp"
+#include "function/aggregate_function/stddev_samp.hpp"
+#include "function/aggregate_function/sum.hpp"
